+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -3,8 +3,8 @@ package kitchenpos.application;
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
 import kitchenpos.request.MenuCreateRequest;
 import kitchenpos.request.MenuProductDto;
