+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -5,6 +5,7 @@ import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.*;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.product.domain.Product;
 import kitchenpos.request.MenuCreateRequest;
 import kitchenpos.request.MenuProductDto;
